Monitor usage class implemented

+++ Models/TotalResourceUsageModel.cs
@@ -5,14 +5,14 @@ namespace RemoteMonitor.Models
     public class TotalResourceUsageModel
     {
 
-        public TotalResourceUsageModel(UInt64 epocTime, float cpuUsage, float memoryUsage)
+        public TotalResourceUsageModel(int epocTime, float cpuUsage, float memoryUsage)
         {
             this.epocTime = epocTime;
             this.cpuUsage = cpuUsage;
             this.memoryUsage = memoryUsage;
         }
         
-        public UInt64 epocTime { get; set; }
+        public int epocTime { get; set; }
 
         public float cpuUsage { get; set; }
